Move `FarmerProtocolInfo` into `subspace-farmer-components`
