+++ crates/subspace-farmer-components/Cargo.toml
@@ -20,14 +20,15 @@ async-trait = "0.1.57"
 bitvec = "1.0.1"
 blake2-rfc = "0.2.18"
 fs2 = "0.4.3"
+hex = { version = "0.4.3", features = ["serde"] }
 libc = "0.2.131"
 parity-scale-codec = "3.1.5"
 rand = "0.8.5"
 schnorrkel = "0.9.1"
+serde = { version = "1.0.143", features = ["derive"] }
 static_assertions = "1.1.0"
 subspace-solving = { version = "0.1.0", path = "../subspace-solving" }
 subspace-core-primitives = { version = "0.1.0", path = "../subspace-core-primitives" }
-subspace-rpc-primitives = { version = "0.1.0", path = "../subspace-rpc-primitives" }
 subspace-verification = { version = "0.1.0", path = "../subspace-verification" }
 thiserror = "1.0.32"
 tracing = "0.1.36"
