+++ crates/subspace-farmer/src/rpc_client/node_rpc_client.rs
@@ -9,8 +9,9 @@ use std::pin::Pin;
 use std::sync::Arc;
 use subspace_archiving::archiver::ArchivedSegment;
 use subspace_core_primitives::{Piece, PieceIndex, RecordsRoot, SegmentIndex};
+use subspace_farmer_components::FarmerProtocolInfo;
 use subspace_rpc_primitives::{
-    FarmerProtocolInfo, RewardSignatureResponse, RewardSigningInfo, SlotInfo, SolutionResponse,
+    RewardSignatureResponse, RewardSigningInfo, SlotInfo, SolutionResponse,
 };
 
 // Defines max_concurrent_requests constant in the node rpc client.
